アカウントID及びユーザー名更新関数を切り出し
