+++ src/app/myPage/components/UserInfo.tsx
@@ -1,16 +1,16 @@
 import React, { useState, useEffect } from 'react'
-import { View, Text, StyleSheet, TouchableOpacity, Alert } from 'react-native'
+import { View, Text, StyleSheet, TouchableOpacity } from 'react-native'
 import { Image } from 'expo-image'
 import { noUserImage } from '../../constants/userImage';
 import EditIcon from '../../components/Icon/EditIcon';
 import { UserInfoType } from '../../../../type/userInfo';
 import UserLogout from '../../actions/handleLogout';
 import UserEditContents from './UserEditContents';
-import { db } from '../../../config';
-import { doc, updateDoc } from 'firebase/firestore'
 import { validateAccountId, validateUserName } from '../../../../utils/validation';
 import Divider from '../../components/Divider';
 import updateUserImage from '../action/backend/updateUserImage';
+import updateAccountId from '../action/backend/updateAccountId';
+import updateUserName from '../action/backend/updateUserName';
 
 type UserInfoProps = {
   userInfos: UserInfoType | null
@@ -51,20 +51,8 @@ export default function UserInfo({ userInfos, userId, userInfoId }: UserInfoProp
   }
 
   // アカウントID更新
-  const handleAccountIdUpdate = async (accountId: string | undefined) => {
-    if (!accountId || !userId || !userInfoId) return;
-    if (errors.accountId) return;
-    try {
-      const userRef = doc(db, `users/${userId}/userInfo/${userInfoId}`);
-      await updateDoc(userRef, {
-        accountId: accountId,
-      });
-      setIsAccountIdEdit(false)
-      Alert.alert("ユーザーIDの更新に成功しました");
-    } catch (error) {
-      console.log("error", error);
-      Alert.alert("ユーザーIDの更新に失敗しました");
-    }
+  const handleUpdateAccountId = async () => {
+    updateAccountId(accountId, errors.accountId, setIsAccountIdEdit, userId, userInfoId);
   }
 
   // ユーザーIDのバリデーション
@@ -73,20 +61,9 @@ export default function UserInfo({ userInfos, userId, userInfoId }: UserInfoProp
     setErrors({ ...errors, accountId: errorMessage })
   }
 
-  const handleUserNameUpdate = async (userUpdateInfo: string | undefined) => {
-    if (!userUpdateInfo || !userId || !userInfoId) return;
-    if (errors.userName) return;
-    try {
-      const userRef = doc(db, `users/${userId}/userInfo/${userInfoId}`);
-      await updateDoc(userRef, {
-        userName: userUpdateInfo,
-      });
-      setIsUserNameEdit(false)
-      Alert.alert("ユーザー名の更新に成功しました");
-    } catch (error) {
-      console.log("error", error);
-      Alert.alert("ユーザー名の更新に失敗しました");
-    }
+  // ユーザー名更新
+  const handleUpdateUserName = async () => {
+    updateUserName(userName, errors.userName, setIsUserNameEdit, userId, userInfoId);
   }
 
   // ユーザー名のバリデーション
@@ -121,7 +98,7 @@ export default function UserInfo({ userInfos, userId, userInfoId }: UserInfoProp
           setIsContentEdit={setIsAccountIdEdit}
           userUpdateContent={accountId}
           setUserUpdateContent={setAccountId}
-          handleUserInfoUpdate={handleAccountIdUpdate}
+          handleUserInfoUpdate={handleUpdateAccountId}
           errorText={errors.accountId}
           handleValidateUserContent={handleValidateAccountId}
         />
@@ -133,7 +110,7 @@ export default function UserInfo({ userInfos, userId, userInfoId }: UserInfoProp
           setIsContentEdit={setIsUserNameEdit}
           userUpdateContent={userName}
           setUserUpdateContent={setUserName}
-          handleUserInfoUpdate={handleUserNameUpdate}
+          handleUserInfoUpdate={handleUpdateUserName}
           errorText={errors.userName}
           handleValidateUserContent={handleValidateUserName}
         />
